+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Add #19: Add the `Yiisoft\Cookies\CookieEncryptor` class to encrypt the value of the cookie and verify that it is tampered (devanych)
 - Add #19: Add the `Yiisoft\Cookies\CookieSigner` class to sign the value of the cookie and verify that it is tampered (devanych)
+- Add #22: Add the `Yiisoft\Cookies\CookieMiddleware` class to encrypt/sign the value of the cookie and verify that it is tampered (devanych)
 
 ## 1.0.0 December 02, 2020